third_party/WebKit/LayoutTests/http/tests/security/link-crossorigin-subresource-no-cors.html

<!DOCTYPE HTML>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Test that a subresource with a crossorigin attribute does not load a cross-origin resource that isn't CORS enabled.");
</script>
<script src="resources/link-crossorigin-common.js"></script>
<link crossorigin="anonymous" rel="subresource" href="http://localhost:8000/security/resources/abe.png?1" onload="fail()" onerror="pass()">
<link crossorigin="use-credentials" rel="subresource" href="http://localhost:8000/security/resources/abe.png?2" onload="fail()" onerror="pass()">
<script>
function runTest()
{
    // Test that dynamically inserted <link> elements are handled the same way.
    var link = document.createElement("link");
    link.rel = "subresource";
    link.crossOrigin = "anonymous";
    link.onload = fail;
    link.onerror = pass;
    link.href = "http://localhost:8000/security/resources/abe.png?3";
    document.body.appendChild(link);

    link = document.createElement("link");
    link.rel = "subresource";
    link.crossOrigin = "use-credentials";
    link.onload = fail;
    link.onerror = pass;
    link.href = "http://localhost:8000/security/resources/abe.png?4";
    document.body.appendChild(link);
}
window.onload = runTest;
</script>
</body>
</html>
